Fixes for map-placed monster corpse decorations
This modifies the DeadDoomImp, DeadDemon, etc. actors.

It does two things:

First, a visual fix - It sets up a system that checks what monster setting the player is using (Original Doom or D4D), and spawns a corpse that's visually appropriate (e.g, it spawns OG imps when playing Original Doom, and Phobos' new ones when playing D4D), so there's no visual inconsistencies. It helps with immersion. :)

Second, when making this, I discovered that if an OG Archvile attempts to raise one of the dead decoration actors, the resulting monster will be unable to move or be hurt. 
I addressed this by sneakily destroying the corpse and replacing it with a brand new actor during the raise state. I suspect there's probably a better way to do it, but my Z-Fu isn't strong enough for that yet.