Doom 4 For Doom Final Commit.
Unless there are some bugs that need fixing, this is it. The last commit, and the last time I'll be touching this mod.

- Removed all references to the Unmaykr and all unimplemented grenades. The Unmaykr will be coming in DE4D instead.
- Reverted the SSG bullet gibbing condition after discovering a nasty side effect of bullets not going through enemies due to the damage delay. SSG now always gibs. Use the shotgun if not desired.
- Rocket/Grenade Launcher now uses a better damage/thrusting protocol which more realistically blasts objects away instead of just straight up.
- Overhauled BFG laser effect.
- Slightly increased the drop chance of upgrades from non-native enemies.
- Moved some more code over to MC_Utility.pk3

A big thank you to everyone past and present who enjoyed this mod. Your support is very much appreciated, and I look forward to seeing you playing DE4D when the time is ripe. ♥